requirements.txt
asgiref==3.2.10
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
autopep8==1.4.4
Babel==2.8.0
boto==2.48.0
boto3==1.4.7
botocore==1.7.32
certifi==2017.7.27.1
cfe==0.0.15
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.2.9
django-3-jet==1.0.8
django-allauth==0.40.0
django-bootstrap-datepicker-plus==3.0.5
django-countries==5.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-feedparser==0.2.1
django-filter==2.3.0
django-jet==1.0.8
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
django-robots==4.0
django-smtp-ssl==1.0
django-storages==1.6.5
django-tinymce4-lite==1.7.5
django-widget-tweaks==1.4.5
docutils==0.14
feedparser==5.2.1
gitdb2==2.0.3
GitPython==2.1.7
google-api-python-client==1.4.1
gunicorn==19.7.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.6
incremental==17.5.0
isort==4.3.21
jmespath==0.9.3
jsmin==2.2.2
lazy-object-proxy==1.4.1
Markdown==3.2.1
markdown2==2.3.8
mccabe==0.6.1
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.44
pep8==1.7.1
phonenumbers==8.11.3
Pillow==4.3.0
pkg-resources==0.0.0
psycopg2==2.7.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.6.1
python-decouple==3.1
python-http-client==3.2.5
python3-openid==3.1.0
pytz==2020.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.11.0
smmap2==2.0.3
sqlparse==0.3.1
static3==0.7.0
txaio==18.8.1
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.3
virtualenv==16.6.1
virtualenvwrapper-win==1.2.5
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.6.0
